Modularizei as definições da api de entities dentro da modulo entities
