+++ main.py
@@ -1,19 +1,14 @@
-from http import HTTPStatus
+from fastapi import FastAPI
+from fastapi.middleware.trustedhost import TrustedHostMiddleware
 
-from bson.errors import InvalidId
-
-from apps.entities.controllers import PersonController
-from apps.entities.schemas import Person
-
+from apps.entities.api import router as entities_router
 from utils.middlewares import ProcessTimeHeader
 from conf.settings import ALLOWED_HOSTS
-
-from fastapi import FastAPI, HTTPException
-from fastapi.middleware.trustedhost import TrustedHostMiddleware
 
 app = FastAPI()
 app.add_middleware(TrustedHostMiddleware, allowed_hosts=ALLOWED_HOSTS)
 app.add_middleware(ProcessTimeHeader)
+app.include_router(entities_router)
 
 
 @app.get("/")
@@ -23,57 +18,6 @@ def root():
         "message": "Object sucessfully returned",
         "object": {}
     }
-
-
-@app.get("/person/", status_code=HTTPStatus.OK)
-async def filter(query: dict | None = None):
-
-    if query is None:
-        query = {}
-
-    try:
-        return PersonController().filter(query)
-
-    except InvalidId:
-        raise HTTPException(status_code=403, detail="Invalid identifier")
-
-
-@app.get("/person/{oid}", status_code=HTTPStatus.OK)
-async def person(oid):
-    try:
-        return PersonController().object(oid)
-
-    except InvalidId:
-        raise HTTPException(status_code=403, detail="Invalid identifier")
-
-
-@app.post("/person/", status_code=HTTPStatus.CREATED)
-async def insert(data: Person):
-    return PersonController().insert(data.dict())
-
-
-@app.put("/person/{oid}", status_code=HTTPStatus.CREATED)
-async def update(oid: str, data: dict):
-    try:
-        return PersonController().update(oid, data)
-
-    except InvalidId:
-        raise HTTPException(status_code=403, detail="Invalid identifier")
-
-
-@app.delete("/person/{oid}", status_code=HTTPStatus.NO_CONTENT)
-async def delete(oid: str):
-    try:
-        deleted = PersonController().delete(oid)
-
-        return {
-            "result": True,
-            "message": "Object deleted successfully",
-            "object": deleted
-        }
-
-    except InvalidId:
-        raise HTTPException(status_code=404, detail="Invalid identifier")
 
 
 if __name__ == "__main__":
